rework of api, logic in revision_logic file now, fails at load_blobs
