+++ handlers/api.py
@@ -12,6 +12,7 @@ import RDF
 
 from models import User, Token, Repo, HMap, CSet, Blob
 from handlers import RequestHandler
+import revision_logic
 
 def authenticated(method):
     """Decorate API methods to require user authentication via token."""
@@ -37,41 +38,7 @@ def now():
 # Pagination size for indexes (number of resource URIs per page)
 INDEX_PAGE_SIZE = 1000
 
-# This factor (among others) determines whether a snapshot is stored rather
-# than a delta, depending on the size of the latest snapshot and subsequent
-# deltas. For the latest snapshot `base` and deltas `d1`, `d2`, ..., `dn`
-# a new snapshot is definitely stored if:
-#
-# `SNAPF * len(base) <= len(d1) + len(d2) + ... + len(dn)`
-#
-# In short, larger values will result in longer delta chains and likely reduce
-# storage size at the expense of higher revision reconstruction costs.
-#
-# TODO: Empirically determine a good value with real data/statistics.
-SNAPF = 10.0
-
 # TODO: Tune zlib compression parameters `level`, `wbits`, `bufsize`?
-
-def compress(s):
-    return zlib.compress(s)
-
-def decompress(s):
-    return zlib.decompress(s)
-
-def shasum(s):
-    return hashlib.sha1(s).digest()
-
-# Parse serialized RDF:
-#
-# RDF/XML:      application/rdf+xml
-# N-Triples:    application/n-triples
-# Turtle:       text/turtle
-def parse(s, fmt):
-    stmts = set()
-    parser = RDF.Parser(mime_type=fmt)
-    for st in parser.parse_string_as_stream(s, "urn:x-default:tailr"):
-        stmts.add(str(st) + " .")
-    return stmts
 
 def join(parts, sep):
     return string.joinfields(parts, sep)
@@ -96,16 +63,17 @@ class BaseHandler(RequestHandler):
 
 class RepoHandler(BaseHandler):
     """Processes repository calls: Push, timegate, memento, timemap etc."""
-
     # def head(self, username, reponame):
     #     pass
 
     def get(self, username, reponame):
+        
         timemap = self.get_query_argument("timemap", "false") == "true"
         index = self.get_query_argument("index", "false") == "true"
         key = self.get_query_argument("key", None)
-
+        
         if (index and timemap) or (index and key) or (timemap and not key):
+        
             raise HTTPError(400)
 
         if self.get_query_argument("datetime", None):
@@ -117,14 +85,9 @@ class RepoHandler(BaseHandler):
         else:
             ts = now()
 
-        try:
-            repo = (Repo
-                .select(Repo.id)
-                .join(User)
-                .where((User.name == username) & (Repo.name == reponame))
-                .naive()
-                .get())
-        except Repo.DoesNotExist:
+        #load repo
+        repo = revision_logic.load_repo(username, reponame)
+        if repo == None:
             raise HTTPError(404)
 
         if key and not timemap:
@@ -135,33 +98,10 @@ class RepoHandler(BaseHandler):
             self.set_header("Content-Type", "application/n-quads")
             self.set_header("Vary", "accept-datetime")
 
-            sha = shasum(key.encode("utf-8"))
+            sha = revision_logic.get_shasum(key)
+            chain = revision_logic.get_chain(repo, sha, ts)
 
-            # Fetch all relevant changes from the last "non-delta" onwards,
-            # ordered by time. The returned delta-chain consists of either:
-            # a snapshot followed by 0 or more deltas, or
-            # a single delete.
-            chain = list(CSet
-                .select(CSet.time, CSet.type)
-                .where(
-                    (CSet.repo == repo) &
-                    (CSet.hkey == sha) &
-                    (CSet.time <= ts) &
-                    (CSet.time >= SQL(
-                        "COALESCE((SELECT time FROM cset "
-                        "WHERE repo_id = %s "
-                        "AND hkey_id = %s "
-                        "AND time <= %s "
-                        "AND type != %s "
-                        "ORDER BY time DESC "
-                        "LIMIT 1), 0)",
-                        repo.id, sha, ts, CSet.DELTA
-                    )))
-                .order_by(CSet.time)
-                .naive())
-
-            if len(chain) == 0:
-                # A resource does not exist for the given key.
+            if chain == None:
                 raise HTTPError(404)
 
             timegate_url = (self.request.protocol + "://" +
@@ -184,54 +124,21 @@ class RepoHandler(BaseHandler):
                 raise HTTPError(404)
 
             # Load the data required in order to restore the resource state.
-            blobs = (Blob
-                .select(Blob.data)
-                .where(
-                    (Blob.repo == repo) &
-                    (Blob.hkey == sha) &
-                    (Blob.time << map(lambda e: e.time, chain)))
-                .order_by(Blob.time)
-                .naive())
-
-            if len(chain) == 1:
-                # Special case, where we can simply return
-                # the blob data of the snapshot.
-                snap = blobs.first().data
-                return self.finish(decompress(snap))
-
-            stmts = set()
-
-            for i, blob in enumerate(blobs.iterator()):
-                data = decompress(blob.data)
-
-                if i == 0:
-                    # Base snapshot for the delta chain
-                    stmts.update(data.splitlines())
-                else:
-                    for line in data.splitlines():
-                        mode, stmt = line[0], line[2:]
-                        if mode == "A":
-                            stmts.add(stmt)
-                        else:
-                            stmts.discard(stmt)
-
-            self.write(join(stmts, "\n"))
+            try:
+                blobs = revision_logic.load_blobs(repo, sha, chain)
+            except ValueError, val:
+                return self.finish(val)
+            
+            self.write(join(blobs, "\n"))
         elif key and timemap:
             # Generate a timemap containing historic change information
             # for the requested key. The timemap is in the default link-format
             # or as JSON (http://mementoweb.org/guide/timemap-json/).
 
-            sha = shasum(key.encode("utf-8"))
-
-            csets = (CSet
-                .select(CSet.time)
-                .where((CSet.repo == repo) & (CSet.hkey == sha))
-                .order_by(CSet.time.desc())
-                .naive())
-
+            sha = revision_logic.get_shasum(key)
+            csit = revision_logic.create_cset(repo, sha)            
+                
             # TODO: Paginate?
-
-            csit = csets.iterator()
 
             try:
                 first = csit.next()
@@ -288,39 +195,16 @@ class RepoHandler(BaseHandler):
 
             page = int(self.get_query_argument("page", "1"))
 
-            # Subquery for selecting max. time per hkey group
-            mx = (CSet
-                .select(CSet.hkey, fn.Max(CSet.time).alias("maxtime"))
-                .where((CSet.repo == repo) & (CSet.time <= ts))
-                .group_by(CSet.hkey)
-                .order_by(CSet.hkey)
-                .paginate(page, INDEX_PAGE_SIZE)
-                .alias("mx"))
+            hm = generate_index(repo, ts, page)
 
-            # Query for all the relevant csets (those with max. time values)
-            cs = (CSet
-                .select(CSet.hkey, CSet.time)
-                .join(mx, on=(
-                    (CSet.hkey == mx.c.hkey_id) &
-                    (CSet.time == mx.c.maxtime)))
-                .where((CSet.repo == repo) & (CSet.type != CSet.DELETE))
-                .alias("cs"))
-
-            # Join with the hmap table to retrieve the plain key values
-            hm = (HMap
-                .select(HMap.val)
-                .join(cs, on=(HMap.sha == cs.c.hkey_id))
-                .naive())
-
-            for h in hm.iterator():
+            for h in hm:
                 self.write(h.val + "\n")
         else:
             raise HTTPError(400)
-
+            
     @authenticated
     def put(self, username, reponame):
         # Create a new revision of the resource specified by `key`.
-
         fmt = self.request.headers.get("Content-Type", "application/n-triples")
         key = self.get_query_argument("key", None)
 
@@ -333,110 +217,28 @@ class RepoHandler(BaseHandler):
         datestr = self.get_query_argument("datetime", None)
         ts = datestr and date(datestr, QSDATEFMT) or now()
 
-        try:
-            repo = (Repo
-                .select(Repo.id)
-                .join(User)
-                .where((User.name == username) & (Repo.name == reponame))
-                .naive()
-                .get())
-        except Repo.DoesNotExist:
+        repo = revision_logic.loadRepo(username, reponame)
+        if repo == None:
             raise HTTPError(404)
 
-        sha = shasum(key.encode("utf-8"))
+        sha = revision_logic.get_shasum(key)
+        chain = revision_logic.get_chain_without_time(repo, sha)
 
-        chain = list(CSet
-            .select(CSet.time, CSet.type, CSet.len)
-            .where(
-                (CSet.repo == repo) &
-                (CSet.hkey == sha) &
-                (CSet.time >= SQL(
-                    "COALESCE((SELECT time FROM cset "
-                    "WHERE repo_id = %s "
-                    "AND hkey_id = %s "
-                    "AND type != %s "
-                    "ORDER BY time DESC "
-                    "LIMIT 1), 0)",
-                    repo.id, sha, CSet.DELTA
-                )))
-            .order_by(CSet.time)
-            .naive())
-
-        if len(chain) > 0 and not ts > chain[-1].time:
-            # Appended timestamps must be monotonically increasing!
+        try:
+            revision_logic.detect_collisions(chain, sha, ts, key)
+        except ValueError:
             raise HTTPError(400)
-
-        if len(chain) == 0:
-            # Mapping for `key` likely does not exist:
-            # Store the SHA-to-KEY mapping in HMap,
-            # looking out for possible collisions.
-            try:
-                HMap.create(sha=sha, val=key)
-            except IntegrityError:
-                val = HMap.select(HMap.val).where(HMap.sha == sha).scalar()
-                if val != key:
-                    raise HTTPError(500)
+        except IntegrityError:
+            raise HTTPError(500)
+        
 
         # Parse and normalize into a set of N-Quad lines
         stmts = parse(self.request.body, fmt)
         snapc = compress(join(stmts, "\n"))
 
-        if len(chain) == 0 or chain[0].type == CSet.DELETE:
-            # Provide dummy value for `patch` which is never stored.
-            # If we get here, we always store a snapshot later on!
-            patch = ""
-        else:
-            # Reconstruct the previous state of the resource
-            prev = set()
-
-            blobs = (Blob
-                .select(Blob.data)
-                .where(
-                    (Blob.repo == repo) &
-                    (Blob.hkey == sha) &
-                    (Blob.time << map(lambda e: e.time, chain)))
-                .order_by(Blob.time)
-                .naive())
-
-            for i, blob in enumerate(blobs.iterator()):
-                data = decompress(blob.data)
-
-                if i == 0:
-                    # Base snapshot for the delta chain
-                    prev.update(data.splitlines())
-                else:
-                    for line in data.splitlines():
-                        mode, stmt = line[0], line[2:]
-                        if mode == "A":
-                            prev.add(stmt)
-                        else:
-                            prev.discard(stmt)
-
-            if stmts == prev:
-                # No changes, nothing to be done. Bail out.
-                return self.finish()
-
-            patch = compress(join(
-                map(lambda s: "D " + s, prev - stmts) +
-                map(lambda s: "A " + s, stmts - prev), "\n"))
-
-        # Calculate the accumulated size of the delta chain including
-        # the (potential) patch from the previous to the pushed state.
-        acclen = reduce(lambda s, e: s + e.len, chain[1:], 0) + len(patch)
-
-        blen = len(chain) > 0 and chain[0].len or 0 # base length
-
-        if (len(chain) == 0 or chain[0].type == CSet.DELETE or
-            len(snapc) <= len(patch) or SNAPF * blen <= acclen):
-            # Store the current state as a new snapshot
-            Blob.create(repo=repo, hkey=sha, time=ts, data=snapc)
-            CSet.create(repo=repo, hkey=sha, time=ts, type=CSet.SNAPSHOT,
-                len=len(snapc))
-        else:
-            # Store a directed delta between the previous and current state
-            Blob.create(repo=repo, hkey=sha, time=ts, data=patch)
-            CSet.create(repo=repo, hkey=sha, time=ts, type=CSet.DELTA,
-                len=len(patch))
+        prev_state = revision_logic.reconstruct_prev_state(chain, repo, sha, stmts, patch, snapc, ts)
+        if prev_state == None:
+            self.finish()
 
     @authenticated
     def delete(self, username, reponame):
@@ -454,38 +256,21 @@ class RepoHandler(BaseHandler):
         datestr = self.get_query_argument("datetime", None)
         ts = datestr and date(datestr, QSDATEFMT) or now()
 
-        try:
-            repo = (Repo
-                .select(Repo.id)
-                .join(User)
-                .where((User.name == username) & (Repo.name == reponame))
-                .naive()
-                .get())
-        except Repo.DoesNotExist:
+        repo = revision_logic.load_repo(username, reponame)
+        if repo == None:
             raise HTTPError(404)
-
-        sha = shasum(key.encode("utf-8"))
-
-        try:
-            last = (CSet
-                .select(CSet.time, CSet.type)
-                .where((CSet.repo == repo) & (CSet.hkey == sha))
-                .order_by(CSet.time.desc())
-                .limit(1)
-                .naive()
-                .get())
-        except CSet.DoesNotExist:
-            # No changeset was found for the given key -
-            # the resource does not exist.
+        
+        sha = revision_logic.get_shasum(key)
+        last = revision_logic.create_last_set(sha, repo)
+        if last[0] == False:
             raise HTTPError(400)
 
-        if not ts > last.time:
+        if not ts > last[1]:
             # Appended timestamps must be monotonically increasing!
             raise HTTPError(400)
 
-        if last.type == CSet.DELETE:
+        if last[2] == CSet.DELETE:
             # The resource was deleted already, return instantly.
             return self.finish()
 
-        # Insert the new "delete" change.
-        CSet.create(repo=repo, hkey=sha, time=ts, type=CSet.DELETE, len=0)
+        insert_delete_changes(repo, last, ts)
